requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
phonenumberslite==8.13.25
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
